Refactoring and added tslint.
