+++ Something/WebApplication/Views/Account/Login.cshtml
@@ -72,7 +72,7 @@
                                     <button type="submit" class="btn btn-default" name="provider" value="@provider.Name" title="Log in using your @provider.DisplayName account">@provider.Name</button>
                                 }
                             </p>
-                        </div>
+                        </div>  
                     </form>
                 }
             }
