+++ Something/WebApplication/package.json
@@ -21,6 +21,7 @@
     "source-map-loader": "^0.2.4",
     "style-loader": "^0.23.0",
     "ts-import-plugin": "^1.5.4",
+    "tslint": "^5.11.0",
     "typescript": "^3.0.3",
     "webpack": "^4.17.1",
     "webpack-cli": "^3.1.0"
