+++ Something/WebApplication/wwwroot/js/.eslintrc.json
@@ -1,0 +1,40 @@
+
+{
+    "env": {
+        "browser": true,
+        "es6": true
+    },
+    "extends": ["eslint:recommended", "airbnb/base", "plugin:react/recommended" ],
+    "parserOptions": {
+        "ecmaFeatures": {
+            "experimentalObjectRestSpread": true,
+            "jsx": true
+        },
+        "sourceType": "module"
+    },
+    "plugins": [
+        "react"
+    ],
+    "rules": {
+        "indent": [
+            "error",
+            4
+        ],
+        "linebreak-style": [
+            "error",
+            "windows"
+        ],
+        "quotes": [
+            "error",
+            "single"
+        ],
+        "semi": [
+            "error",
+            "always"
+        ],
+        "prefer-destructuring": "off",
+        "class-methods-use-this": "off",
+        "no-useless-escape": "off",
+        "no-mixed-requires": "off"
+    }
+}
