+++ Something/WebApplication/wwwroot/js/apps/login/app.scss
@@ -1,0 +1,17 @@
+.authentication {
+    .ant-input {
+        min-height: 10px !important;
+    }
+    
+    .pane {
+        height: 100%;
+        padding: 10% 10px 10px 0px;
+    }
+    .login, .register {
+        width: 400px;
+        margin: 0 auto;
+        h2 {
+            font-size: 32px;
+        }
+    }
+}
