+++ Something/WebApplication/wwwroot/js/apps/login/app.tsx
@@ -3,8 +3,8 @@ import * as React from 'react';
 import './../global.scss';
 import './login-app.scss';
 
-import { LoginForm } from './components/LoginForm';
-import { Splash } from './components/Splash'
+import { LoginForm } from './components/LoginForm/LoginForm';
+import { Splash } from './components/Splash/Splash'
 
 export default class Login extends React.Component<{}, {}> {
     render() {
