+++ Something/WebApplication/wwwroot/js/apps/login/components/LoginForm/LoginForm.tsx
@@ -2,6 +2,8 @@ import * as React from 'react';
 import { Form, Input, Button } from 'antd';
 import { FormComponentProps } from 'antd/lib/form'
 
+import './LoginForm.scss';
+
 const FormItem = Form.Item;
 
 interface LoginProps {
