--- Something/WebApplication/wwwroot/js/login-app/login-app.scss
@@ -1,31 +1,0 @@
-.authentication {
-    .ant-input {
-        min-height: 10px !important;
-    }
-    .splash {
-        background: linear-gradient(to bottom, #2f54eb 0%, #3a40d4 100%);
-        color: #fff;
-        &--inner {
-            width: 400px;
-            margin: 0 auto;
-        }
-        &--header {
-            font-size: 32px;
-        }
-        img {
-            margin-top: 30px;
-            width: 400px;
-        }
-    }
-    .pane {
-        height: 100%;
-        padding: 10% 10px 10px 0px;
-    }
-    .login, .register {
-        width: 400px;
-        margin: 0 auto;
-        h2 {
-            font-size: 32px;
-        }
-    }
-}
