+++ Something/WebApplication/wwwroot/js/tsconfig.json
